refactor: extract models prompting cli
